Added scripts section to YAML file and created scripts directory
All scripts that will be provided for the client will need to be located in the new "scripts" directory.

The "use-all" option determines if every single script in the scripts directory will be copied to FTP at run time.

If "use-all" is false, only the scripts specified in the "scripts-list" list in the YAML file will be copied to the FTP directory.